refactor(gateway): extract FederationUpstreamConfig (GW-2 S2)
